refactor: use iterators where possible in Game impl

+++ src/lib.rs
@@ -7,13 +7,13 @@ pub mod game;
 pub mod graphics;
 pub mod logger;
 
-pub const WIDTH: i32 = 10;
-pub const HEIGHT: i32 = 10;
+pub const WIDTH: usize = 10;
+pub const HEIGHT: usize = 10;
 
 pub const WIDTH_PX: i32 = 800;
 pub const HEIGHT_PX: i32 = 800;
 
-pub const MINE_COUNT: i32 = WIDTH * HEIGHT / 5;
+pub const MINE_COUNT: usize = WIDTH * HEIGHT / 5;
 
 pub fn clear_draw(r: f32, g: f32, b: f32, a: f32) {
     unsafe {
